refactor: only use Stream for the permutations, not recursion
Now that we can have tests of the worst-case behavior without needing to
fake recursing through the entire stream, we can go back to a
slightly-simpler implementation. We use the stream to generate permutations,
but a normal recursion for the employees.

We also change the `permutations_take` function so that it modifies the
accumulator as a part of its recursion, rather than a separate second step.