feat[EmployeeVacationQuotaSummary]: use maximum_minutes and full-year vacation
`maximum_minutes` takes FMLA leave into account, so it reduced the amount of
vacation that it's possible to take even without assigned vacation.

Vacation across the entire year counts against the operator's quota, so we
calculate the full year (based on the end date, so as to handle the
Winter/Annual pick) and find any vacation booked in that interval.